feat: ts api jinja generator (#250)
* feat: ts api jinja generator

* refactor: addressing pr comments

* refactor: code cleanup, docstrings; temporary duplicate of SignedTransaction from utils

* refactor: wip ensure kebab-case names for generated models in for TS clients

* chore: tmp remove artifacts

* chore: remove root artifacts from ts template (will be managed manually)

* refactor: generating packages into utils ts npm workspace; switching bun to npm

* chore: wip

* chore: adding dto generation into model files

* refactor: add transformation to strip json responses from select endpoints to align with go and js sdks

* chore: wip

* refactor: refining ts generator gh actions; mirroring rust tests structure

* refactor: finalizing integration of algod and indexer tests into utils ts package

* ci: fixing package resolution for ts api ci

* chore: restore readme; further tweaks for ci pipeline

* chore: fix wait for indexer test order

* chore: pr comments

+++ crates/indexer_client/src/models/transaction.rs
@@ -79,13 +79,13 @@ pub struct Transaction {
         rename = "created-application-index",
         skip_serializing_if = "Option::is_none"
     )]
-    pub created_application_index: Option<u32>,
+    pub created_application_index: Option<u64>,
     /// Specifies an asset index (ID) if an asset was created with this transaction.
     #[serde(
         rename = "created-asset-index",
         skip_serializing_if = "Option::is_none"
     )]
-    pub created_asset_index: Option<u32>,
+    pub created_asset_index: Option<u64>,
     /// \[fee\] Transaction fee.
     #[serde(rename = "fee")]
     pub fee: u64,
